Fehler jump to case

+++ include/hashMacStorage.h
@@ -8,11 +8,11 @@
 
 void visitor_mac_add(uint16_t hashedmac, int8_t nextPos);
 void visitor_ausgabe();
+
 void clear_storage(hashedMacBuffer_t storage);
 
 //return fuer senddata 
-int return_visitor_mac(int pos);
-
+uint16_t return_visitor_mac(int pos);
 
 
 #endif
